requirements.txt
# LLM & Agent Framework
langchain==0.3.13
langchain-ollama==0.2.0
langchain-community==0.3.13
langchain-core==0.3.28

# Ollama Integration
ollama==0.4.4

# Data Processing
pandas==2.2.3
numpy<2.0,>=1.22.4

# UI
gradio==5.9.1

# Environment Variables
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-cov==6.0.0

# Utilities
pydantic==2.10.5
httpx>=0.27.0